requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Werkzeug==3.1.3
anthropic==0.42.0
python-dotenv==1.0.1
gunicorn==23.0.0
